Styles object
A css function was created that converted the styles object to a string. It works like this: first, using the bject.keys method, the keys of the dictionary were extracted. Then, using the map () method, the obtained keys and their objects were glued into a string. The join (';') method was used to correctly split strings.
In order for the resulting methods to be processed correctly, they were substituted into the code collector in the row () function.