toHTML function
The toHTML() function was created in the parent class Block, and it throws an error about the absence of a function in the child class.
In the TitleBlock, TextBlock, ColumnsBlock, ImageBlock classes, the toHTML() function executes the whole function from the templates.js file. hence all functionality of templates.js file goes into classes of blocks.js file. The templates.js file has been removed.