Create class SiteBar
A SiteBar class has been created with a constructor that creates a document element with a call to the querySelector method of a record on the required #panel element from html.
In addition, the block () function has been created, which creates a field for filling #panel.
This function is used in the init () method, which ends the execution of the constructor. The block () function takes an instance of an object which will accept