Adding collected items to the page
In the index.js file, the SiteBar object, in addition to the parameter to which it belongs, added the updateCallback function. This function first adds a new block newBlock to the end of the model, and then inserts the resulting result into the html code.
When the updateCallback function or update method is run, a poll is started from the 'submit' element, where, in turn, the event object polls the value of the value variable and writes them to the value and styles variables.
Further, when choosing which field was queried, the code is collected using the TextBlock TitleBlock classes, the result is written to the newBlock variable, which is the argument of the update method.