Program start
An App class was created that took a model object in its constructor.
All site assembly and monitoring behind the control panel have been moved from the index.js file to the init () method of the App class.
In the index.js file, when the App class object is created, the application is launched.